remove some useless dep
